refactor: ListViewUser를 ListRow로 리팩터링
