+++ frontend/src/pages/SelectReceiver/components/ListViewUsers.tsx
@@ -1,7 +1,9 @@
 import styled from '@emotion/styled';
+import Avatar from '../../../components/Avatar';
+import { ListRow } from '../../../components/ListRow';
 import { UserProfile } from '../../../types/user';
 import useListViewUsers from '../hooks/useListViewUsers';
-import ListViewUser from './ListViewUser';
+import CheckIcon from '@mui/icons-material/Check';
 
 const ListViewUsers = ({
   onClickUser,
@@ -17,11 +19,23 @@ const ListViewUsers = ({
   return (
     <S.Container>
       {matchedUsers.map(user => (
-        <ListViewUser
-          key={user.id}
-          user={user}
-          onClickUser={onClickUser}
-          isCheckedUser={isCheckedUser}
+        <ListRow
+          left={<Avatar src={user.imageUrl} alt={user.name} />}
+          center={
+            <ListRow.Text2Rows
+              top={user.name}
+              topProps={{ color: 'white', fontSize: '1.5rem' }}
+              bottom={user.email}
+              bottomProps={{ color: '#8e8e8e' }}
+            />
+          }
+          right={
+            <S.CheckBox isChecked={isCheckedUser(user)}>
+              <S.CheckboxIcon />
+            </S.CheckBox>
+          }
+          css={{ color: 'white', cursor: 'pointer' }}
+          onClick={() => onClickUser(user)}
         />
       ))}
     </S.Container>
@@ -29,6 +43,8 @@ const ListViewUsers = ({
 };
 
 export default ListViewUsers;
+
+type CheckBoxProp = { isChecked: boolean };
 
 const S = {
   Container: styled.div`
@@ -60,5 +76,18 @@ const S = {
         border-radius: 5px;
       }
     }
+  `,
+  CheckBox: styled.div<CheckBoxProp>`
+    justify-self: end;
+    margin-right: 5px;
+    display: ${({ isChecked }) => (isChecked ? 'flex' : 'none')};
+    align-items: center;
+  `,
+  CheckboxIcon: styled(CheckIcon)`
+    border-radius: 50%;
+    background-color: ${({ theme }) => theme.primary};
+    border: 1px solid black;
+    fill: white;
+    padding: 1px;
   `,
 };
